+++ src/model/points-model.js
@@ -2,7 +2,7 @@ import {getRandomPoint} from '../mock/points.js';
 import { mockOffers } from '../mock/offers.js';
 import { mockDestinations } from '../mock/destinations.js';
 
-const POINT_COUNT = 4;
+const POINT_COUNT = 10;
 
 export default class PointsModel {
   #points = Array.from({length: POINT_COUNT}, getRandomPoint);
