+++ src/view/create-point-view.js
@@ -1,6 +1,9 @@
 import AbstractView from '../framework/view/abstract-view.js';
 import { POINT_TYPES, DESTINATIONS, DATETIME_FORMAT_FOR_EDIT_FORM } from '../const.js';
-import { humanizeDate, getFormattedType, getOffersForPoint, getDestinationForPoint } from '../utils.js';
+import { humanizeDate } from '../utils/date.js';
+import { getFormattedType } from '../utils/common.js';
+import {getOffersForPoint} from '../utils/point.js';
+import {getDestinationForPoint} from '../utils/point.js';
 
 function createAddPointTypePointTemplate () {
   return POINT_TYPES.map((type)=> `<div class="event__type-item">
